ORION-28: added a guard page to protect against (or actually, catch) stack overflow in fibers.
This will work in debug mode only. This adds an additional protection in addition to the existing one (MAGIC), which works in release mode too. The MAGIC is limited to the case in which we overwrite the begining of the stack, which might be skipped (hence adding a page guard). The unit test includes also a test for the (previously implemented) MAGIC.