ORION-45: scheduler improvements.
Added the ability to wake up a sleeping fiber, i.e. remove it from the sleep queue and return it to the ready queue.

The vmsg and nfs providers will sleep (for 1 msec) if the system is idle (i.e. if they don't receive requests for 5 msec). The vmsg provider will only go to sleep if it has no pending messages (waiting response).
The vmsg provider will wake up if a request is sent (so that it will poll for its response).

The io provider will have a different mechanism. It has no notion of idleness, because it's not listening on incoming messages - it's receiving requests from our fibers. So it will only run if there are active devices. Otherwise, it will be suspended.

Other (unrelated) changes:
1. Changed the TcpAcceptor polling so that instead of a zero timeout it will block while waiting for new connections.
2. Changed the (trace) dumper sleep from 10 micro to 1 msec.